requirements-test.txt
coverage==3.7.1
factory-boy==2.4.1
django>=1.5.1
django-tastypie>=0.12.0
django-guardian>=1.2.4